added prep_chart_data_for_hichart
