+++ app/views/admin/auditor/report_layout.html.erb
@@ -77,6 +77,7 @@
 	<script src="//ajax.googleapis.com/ajax/libs/jquery/1.11.0/jquery.min.js"></script>
 	<script src="//maxcdn.bootstrapcdn.com/bootstrap/3.1.1/js/bootstrap.min.js"></script>
 	<script>
+		var chart_data = <%= @chart_data.to_json.html_safe %>;
 	</script>
 	<%= javascript_include_tag "dashboard" %>
 	<script src="https://code.highcharts.com/highcharts.js"></script>
